[ERM-5124] тип свойств, прописанных в DoubleGis.Erm.Platform.Model.Entities.Enums.Infrastructure.PropertiesToEnumMapper, изменен на соответствующие enum'ы
git-tfs-id: [http://uk-tfs02:8080/tfs/DefaultCollection]$/Erm.Core/Dev/Features/EFCF/BLCore;C25688

+++ Source/API/2Gis.Erm.BLCore.API.Aggregates/Orders/ReadModel/OrderSpecs.Bargains.cs
@@ -30,12 +30,12 @@ namespace DoubleGis.Erm.BLCore.API.Aggregates.Orders.ReadModel
 
                 public static FindSpecification<Bargain> ClientBargains()
                 {
-                    return new FindSpecification<Bargain>(x => x.BargainKind == (int)BargainKind.Client);
+                    return new FindSpecification<Bargain>(x => x.BargainKind == BargainKind.Client);
                 }
 
                 public static FindSpecification<Bargain> AgentBargains()
                 {
-                    return new FindSpecification<Bargain>(x => x.BargainKind == (int)BargainKind.Agent);
+                    return new FindSpecification<Bargain>(x => x.BargainKind == BargainKind.Agent);
                 }
 
                 public static FindSpecification<Bargain> Duplicate(long bargainId,
